+++ internal/provider/provider_test.go
@@ -1,22 +1,14 @@
 package provider
 
 import (
-	"testing"
-
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	"github.com/hashicorp/terraform-plugin-framework/providerserver"
+	"github.com/hashicorp/terraform-plugin-go/tfprotov6"
 )
 
-// ProviderFactories are used to instantiate a provider during acceptance testing.
-// The factory function will be invoked for every Terraform CLI command executed
-// to create a provider server to which the CLI can reattach.
-var ProviderFactories = map[string]func() (*schema.Provider, error){
-	"cue": func() (*schema.Provider, error) {
-		return New("dev")(), nil
-	},
-}
-
-func TestProvider(t *testing.T) {
-	if err := New("dev")().InternalValidate(); err != nil {
-		t.Fatalf("err: %s", err)
-	}
+// testProtoV6ProviderFactories are used to instantiate a provider during
+// acceptance testing. The factory function will be invoked for every Terraform
+// CLI command executed to create a provider server to which the CLI can
+// reattach.
+var testProtoV6ProviderFactories = map[string]func() (tfprotov6.ProviderServer, error){
+	"cue": providerserver.NewProtocol6WithError(New("test")()),
 }