+++ CHANGELOG.md
@@ -1,1 +1,12 @@
+## 0.2.0 (Upcoming)
+    
+CHANGES:
+
+* Provider has been migrated from `terraform-plugin-sdkv2` to
+  `terraform-plugin-framework`.
+
+* Attributes `expression`, `files` and `working_dir` inside **cue_export** data
+  source have been renamed to `path`, `args` and `dir` to better reflect the
+  CUE terminology.
+
 ## 0.1.0 (Initial Release)